segmenter/requirements.txt
Flask>=2.0
pymongo>=4.0
tensorflow==2.14.0
Pillow>=9.0
opencv-python-headless>=4.5
